Move sstableloader upgrade tests to upgrade_tests/storage_engine_upgrade_test.py
